Remove deprecated APIs
